requirements.txt
youtube-transcript-api
yt-dlp
openai
chromadb
streamlit
python-dotenv
